lab2/user_app/requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.1.3